minor styling updates and renaming
