+++ apps/consent-ui/src/app/[lang]/page.tsx
@@ -18,7 +18,7 @@
  */
 
 import { ValidLanguage } from '@/i18n';
-import HomeComponent from '@/components/Home/Home';
+import HomeComponent from '@/components/Home';
 
 export default async function Home({ params: { lang } }: { params: { lang: ValidLanguage } }) {
 	return <HomeComponent currentLang={lang} />;
